Start making cropping and OCR function abstract
The tests have all worked very well for the OWMatchResult class, but it's getting a little messy, and the resulting dictionary isn't logically matched to the screenshot. This is due to what I chose to reduce the screen capture areas to. I'm going to make that class obsolete by doing the following:
1) Make the scree cropping and OCR scanning process abstract and reusable for any project.
2) Make a hierarchy of how I'd like to see the resulting dictionary, then pass that into the class mentioned in step 1.
3) After it the OCR has run in step 2, then do custom formatting on the results.

This refactor will take a moment, but the result will be much more legible code with the added benefit of being somewhat reusable, which is insanely helpful given what we're dealing with here; OCR.