Directory Options
jsonUtil.py:
A new utility file for reading from or writing to a json file. Dictionary in, dictionary out.

osUtil.py:
Only has one function so far, which is to get the most recent file from a folder. String in string out.

tkUtil.py:
Added default parameters for SelectDirectory; now the helper function can be loaded with a default folder.

ScreenshotPath.json: A result of the new code.

main.py:
The program now prompts the user to pick a folder where there OW screenshots are located, if they haven't picked a folder already. In debug mode, the screenshots will be pulled from TestData.
Whenever the program is run, it uses the newest screenshot, chronologically.